+++ packages/overlay-rc/src/use-overlay-manager.tsx
@@ -3,8 +3,8 @@ import { nanoid } from 'nanoid';
 import type { OverlayData, OverlayOptions } from './types.ts';
 import { awaitIfPromise } from './awaitIfPromise';
 
-// 전역 Overlay 상태 관리 (useBeforeClose에서 사용하므로 export)
-export const overlays = signal<OverlayData<any, any>[]>([]);
+// 전역 Overlay 상태 관리
+export const overlays = signal<Array<OverlayData<any, any>>>([]);
 
 // 활성화된 Overlay 계산
 const activeOverlays = computed(() =>
@@ -16,23 +16,10 @@ export const useOverlayManager = () => {
   /* Overlay 닫기 처리 함수
   여기 함수에서는 option이 push 되지 않았기 때문에 
   */
-  const closeOverlay = <TResult,>(
-    id: string,
-    resolve: (value: TResult | undefined) => void, // 타입 명시
-    result?: TResult,
+  const closeOverlay = (
+    id: string
   ) => {
-    const overlayIndex = overlays.value.findIndex((o) => o.id === id);
-    if (overlayIndex > -1) {
-      // overlays 배열을 직접 수정하지 않고, 새로운 배열을 생성하여 업데이트
-      overlays.value = [
-        ...overlays.value.slice(0, overlayIndex),
-        ...overlays.value.slice(overlayIndex + 1),
-      ];
-
-      resolve(result); // close 함수 호출 시 resolve 호출
-    } else {
-      resolve(undefined); // Overlay가 없으면 undefined resolve
-    }
+    overlays.value = overlays.value.filter((overlay) => overlay.id !== id);
   };
 
   // Overlay 열기 함수
@@ -56,13 +43,13 @@ export const useOverlayManager = () => {
 
           // 닫기가 취소된 경우
           if (!canClose) {
-            resolve(undefined); // Promise를 undefined로 resolve
             return false; // 추가 작업 중단
           }
 
           // 기존 오버레이를 닫고 Promise resolve
-          closeOverlay(id, resolve);
-          void options.onClose?.(undefined);
+          closeOverlay(id);
+          void existingOverlay.onClose?.(undefined);
+          resolve(undefined); // Promise를 undefined로 resolve
         }
         return true;
       };
@@ -74,8 +61,9 @@ export const useOverlayManager = () => {
           id,
           open: true,
           onClose: (result) => {
-            closeOverlay(id, resolve, result);
+            closeOverlay(id);
             void options.onClose?.(result);
+            resolve(result);
           },
         };
 
@@ -107,7 +95,7 @@ export const useOverlayManager = () => {
         return; // 닫기 취소
       }
 
-      closeOverlay(id, () => undefined);
+      closeOverlay(id);
       void overlay.onClose?.(undefined);
     }
   };